// src/api/openapi/auto-generated/models/FileResponse.ts
export interface FileResponse {
    /**
     * 
     * @type {string}
     * @memberof FileResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof FileResponse
     */
    filename: string;
    /**
     * 
     * @type {FilePurpose}
     * @memberof FileResponse
     */
    purpose: FilePurpose;
    /**
     * 
     * @type {JupiterApiFileSchemasFileType}
     * @memberof FileResponse
     */
    type: JupiterApiFileSchemasFileType;
    /**
     * 
     * @type {number}
     * @memberof FileResponse
     */
    size: number;
    /**
     * 
     * @type {Date}
     * @memberof FileResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof FileResponse
     */
    updatedAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof FileResponse
     */
    expiresAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof FileResponse
     */
    createdBy: string;
    /**
     * 
     * @type {string}
     * @memberof FileResponse
     */
    updatedBy: string;
    /**
     * 
     * @type {Metadata}
     * @memberof FileResponse
     */
    metadata?: Metadata;
}