// src/api/openapi/auto-generated/models/FileValidationStatus.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const FileValidationStatus = {
    pending: 'pending',
    partiallySucceeded: 'partially_succeeded',
    succeeded: 'succeeded',
    failed: 'failed',
    unknownDefaultOpenApi: '11184809'
} as const;
export type FileValidationStatus = typeof FileValidationStatus[keyof typeof FileValidationStatus];


export function instanceOfFileValidationStatus(value: any): boolean {
    for (const key in FileValidationStatus) {
        if (Object.prototype.hasOwnProperty.call(FileValidationStatus, key)) {
            if ((FileValidationStatus as Record<string, FileValidationStatus>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function FileValidationStatusFromJSON(json: any): FileValidationStatus {
    return FileValidationStatusFromJSONTyped(json, false);
}

export function FileValidationStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): FileValidationStatus {
    return json as FileValidationStatus;
}

export function FileValidationStatusToJSON(value?: FileValidationStatus | null): any {
    return value as any;
}