// src/api/openapi/auto-generated/models/GeocodeLocationMessageStatus.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const GeocodeLocationMessageStatus = {
    passed: 'passed',
    failed: 'failed',
    unknownDefaultOpenApi: '11184809'
} as const;
export type GeocodeLocationMessageStatus = typeof GeocodeLocationMessageStatus[keyof typeof GeocodeLocationMessageStatus];


export function instanceOfGeocodeLocationMessageStatus(value: any): boolean {
    for (const key in GeocodeLocationMessageStatus) {
        if (Object.prototype.hasOwnProperty.call(GeocodeLocationMessageStatus, key)) {
            if ((GeocodeLocationMessageStatus as Record<string, GeocodeLocationMessageStatus>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function GeocodeLocationMessageStatusFromJSON(json: any): GeocodeLocationMessageStatus {
    return GeocodeLocationMessageStatusFromJSONTyped(json, false);
}

export function GeocodeLocationMessageStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): GeocodeLocationMessageStatus {
    return json as GeocodeLocationMessageStatus;
}

export function GeocodeLocationMessageStatusToJSON(value?: GeocodeLocationMessageStatus | null): any {
    return value as any;
}