// src/api/openapi/auto-generated/models/GeocodeValidationStatus.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const GeocodeValidationStatus = {
    pending: 'pending',
    partiallySucceeded: 'partially_succeeded',
    succeeded: 'succeeded',
    failed: 'failed',
    unknownDefaultOpenApi: '11184809'
} as const;
export type GeocodeValidationStatus = typeof GeocodeValidationStatus[keyof typeof GeocodeValidationStatus];


export function instanceOfGeocodeValidationStatus(value: any): boolean {
    for (const key in GeocodeValidationStatus) {
        if (Object.prototype.hasOwnProperty.call(GeocodeValidationStatus, key)) {
            if ((GeocodeValidationStatus as Record<string, GeocodeValidationStatus>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function GeocodeValidationStatusFromJSON(json: any): GeocodeValidationStatus {
    return GeocodeValidationStatusFromJSONTyped(json, false);
}

export function GeocodeValidationStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): GeocodeValidationStatus {
    return json as GeocodeValidationStatus;
}

export function GeocodeValidationStatusToJSON(value?: GeocodeValidationStatus | null): any {
    return value as any;
}