// src/api/openapi/auto-generated/models/LocationGeocodeResponse.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GeocodeLocationIndividualResult } from './GeocodeLocationIndividualResult';
import {
    GeocodeLocationIndividualResultFromJSON,
    GeocodeLocationIndividualResultFromJSONTyped,
    GeocodeLocationIndividualResultToJSON,
} from './GeocodeLocationIndividualResult';
import type { GeocodeValidationStatus } from './GeocodeValidationStatus';
import {
    GeocodeValidationStatusFromJSON,
    GeocodeValidationStatusFromJSONTyped,
    GeocodeValidationStatusToJSON,
} from './GeocodeValidationStatus';

/**
 * 
 * @export
 * @interface LocationGeocodeResponse
 */
export interface LocationGeocodeResponse {
    /**
     * 
     * @type {GeocodeValidationStatus}
     * @memberof LocationGeocodeResponse
     */
    overallStatus: GeocodeValidationStatus;
    /**
     * 
     * @type {Array<GeocodeLocationIndividualResult>}
     * @memberof LocationGeocodeResponse
     */
    individualResults: Array<GeocodeLocationIndividualResult> | null;
}

/**
 * Check if a given object implements the LocationGeocodeResponse interface.
 */
export function instanceOfLocationGeocodeResponse(value: object): value is LocationGeocodeResponse {
    if (!('overallStatus' in value) || value['overallStatus'] === undefined) return false;
    if (!('individualResults' in value) || value['individualResults'] === undefined) return false;
    return true;
}

export function LocationGeocodeResponseFromJSON(json: any): LocationGeocodeResponse {
    return LocationGeocodeResponseFromJSONTyped(json, false);
}

export function LocationGeocodeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationGeocodeResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'overallStatus': GeocodeValidationStatusFromJSON(json['overall_status']),
        'individualResults': (json['individual_results'] == null ? null : (json['individual_results'] as Array<any>).map(GeocodeLocationIndividualResultFromJSON)),
    };
}

export function LocationGeocodeResponseToJSON(value?: LocationGeocodeResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'overall_status': GeocodeValidationStatusToJSON(value['overallStatus']),
        'individual_results': (value['individualResults'] == null ? null : (value['individualResults'] as Array<any>).map(GeocodeLocationIndividualResultToJSON)),
    };
}